requirements.txt
cryptography==42.0.2
django-ninja==1.1.0
email-validator==2.1.0.post1
pip-chill==1.0.3
pyjwt==2.8.0
python-decouple==3.8
